RatioAdapter moves to the proper place
