+++ robozonky-api/src/main/java/com/github/robozonky/api/Ratio.java
@@ -16,12 +16,15 @@
 
 package com.github.robozonky.api;
 
-import java.math.BigDecimal;
-import java.util.Objects;
-
 import static com.github.robozonky.internal.util.BigDecimalCalculator.times;
 import static com.github.robozonky.internal.util.BigDecimalCalculator.toScale;
 
+import java.math.BigDecimal;
+import java.util.Objects;
+
+import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
+
+@XmlJavaTypeAdapter(RatioAdapter.class)
 public final class Ratio extends Number implements Comparable<Ratio> {
 
     public static final Ratio ZERO = new Ratio(BigDecimal.ZERO);
