+++ robozonky-api/src/main/java/com/github/robozonky/api/remote/entities/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 The RoboZonky Project
+ * Copyright 2020 The RoboZonky Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,14 +19,12 @@
         @XmlJavaTypeAdapter(type = OffsetDateTime.class, value = OffsetDateTimeAdapter.class),
         @XmlJavaTypeAdapter(type = LocalDate.class, value = LocalDateAdapter.class),
         @XmlJavaTypeAdapter(type = YearMonth.class, value = YearMonthAdapter.class),
-        @XmlJavaTypeAdapter(type = Ratio.class, value = RatioAdapter.class)
 })
 package com.github.robozonky.api.remote.entities;
 
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
 import java.time.YearMonth;
+
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapters;
-
-import com.github.robozonky.api.Ratio;
